+++ Editor/Unity.Build.Classic.Private/BuildContextExtensions.cs
@@ -20,6 +20,11 @@ namespace Unity.Build.Classic.Private
         };
 
         public static bool UsesIL2CPP(this BuildContext context) => context.TryGetComponent(out Unity.Build.Classic.ClassicScriptingSettings css) && css.ScriptingBackend == ScriptingImplementation.IL2CPP;
-        public static bool IsDevelopmentBuild(this BuildContext context) => context.TryGetComponent(out Unity.Build.Classic.ClassicBuildProfile classicBuildProfile) && classicBuildProfile.Configuration == BuildType.Develop;
+        public static bool IsDevelopmentBuild(this BuildContext context)
+        {
+            if (context.TryGetComponent(out ClassicBuildProfile classicBuildProfile))
+                return classicBuildProfile.Configuration == BuildType.Develop || classicBuildProfile.Configuration == BuildType.Debug;
+            return false;
+        }
     }
 }